+++ route/src/main/java/sparta/rooibos/route/presentation/controller/RouteController.java
@@ -8,7 +8,7 @@ import sparta.rooibos.route.presentation.dto.request.GetOptimizedRouteRequest;
 import sparta.rooibos.route.presentation.dto.request.SearchRouteRequest;
 import sparta.rooibos.route.presentation.dto.request.UpdateRouteRequest;
 import sparta.rooibos.route.presentation.dto.response.*;
-import sparta.rooibos.route.service.RouteService;
+import sparta.rooibos.route.service.port.in.RouteService;
 
 import java.util.UUID;
